requirements.txt
# LangChain core
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-chroma>=0.1.0

# OpenAI
openai>=1.30.0

# Vector store
chromadb>=0.5.0

# Document loaders
pypdf>=4.0.0
python-docx>=1.1.0

# Configuration
python-dotenv>=1.0.0

# CLI & display
typer>=0.12.0
rich>=13.7.0

# Testing
pytest>=8.0.0
pytest-mock>=3.12.0
